agent-core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentarium-core"
version = "0.1.0"
description = "Agentarium agent-core — 自律 AI エージェントのコアエンジン"
requires-python = ">=3.12"
dependencies = [
    # HTTP
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    # RSS
    "feedparser>=6.0.11",
    # HTML 解析
    "beautifulsoup4>=4.12.3",
    # Qdrant
    "qdrant-client>=1.12.0",
    # YAML
    "pyyaml>=6.0.2",
    # Playwright (rebrowser)
    "rebrowser-playwright>=1.48.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "respx>=0.21.0",  # httpx モック
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
